:pencil2: renaming and comments for causal flow codes
